Moved psutil dependency to tests only
