+++ UnitTests/test_ResultTables.py
@@ -8,10 +8,11 @@ PROJECT_ROOT = os.path.abspath(os.path.join(
                   os.pardir)
 )
 sys.path.append(PROJECT_ROOT)
-from RFEM.initModel import Model, getPathToRunningRFEM
+from RFEM.initModel import Model
 from RFEM.connectionGlobals import url
 from RFEM.enums import CaseObjectType
 from RFEM.Results.resultTables import ResultTables
+from tools import getPathToRunningRFEM
 
 if Model.clientModel is None:
     Model()
